+++ src/main/java/cart/config/WebMvcConfig.java
@@ -1,7 +1,6 @@
-package cart;
+package cart.config;
 
 import cart.dao.MemberDao;
-import cart.ui.MemberArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
