+++ src/main/java/cart/ui/view/PageController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.view;
 
 import cart.application.ProductService;
 import cart.dao.MemberDao;
